+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Added <podcast:locked> support.
 - Added <podcast:funding> support.
 - Added <podcast:season> support.
+- Added <podcast:license> support.
 
 ## 0.7.0 - 2023-05-23